+++ grand/analysis/energy_reco/__init__.py
@@ -1,0 +1,5 @@
+"""Energy reconstruction."""
+
+from .voltage import (recons_energy_from_voltage)
+
+__all__ = ['recons_energy_from_voltage']