+++ grand/recon/__init__.py
@@ -1,1 +1,0 @@
-"""Reconstruction of shower parameters from recorded signals."""